+++ Tests/PowerSyncTests/Kotlin/TestLogger.swift
@@ -1,0 +1,11 @@
+@testable import PowerSync
+
+
+class TestLogWriterAdapter: LogWriterProtocol {
+    var logs = [String]()
+    
+    func log(severity: LogSeverity, message: String, tag: String?) {
+        logs.append("\(severity): \(message) \(tag != nil ? "\(tag!)" : "")")
+    }
+}
+